refactor: rename all 'carousels' under widgets to 'tab'
